🛠️ :: compose component 이동
