+++ presentation/src/main/java/team/retum/jobis_android/feature/company/CompaniesScreen.kt
@@ -29,7 +29,7 @@ import com.jobis.jobis_android.R
 import team.retum.domain.entity.company.CompanyEntity
 import team.retum.jobis_android.feature.home.ApplyCompaniesItemShape
 import team.retum.jobis_android.feature.recruitment.Filter
-import team.retum.jobis_android.feature.recruitment.Header
+import team.retum.jobis_android.util.compose.component.Header
 import team.retum.jobis_android.viewmodel.company.CompanyViewModel
 import team.retum.jobisui.colors.JobisColor
 import team.returm.jobisdesignsystem.image.JobisImage
