+++ presentation/src/main/java/team/retum/jobis_android/util/compose/component/Header.kt
@@ -1,0 +1,37 @@
+package team.retum.jobis_android.util.compose.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material.Divider
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import team.retum.jobisui.colors.JobisColor
+import team.returm.jobisdesignsystem.theme.Body2
+
+@Composable
+internal fun Header(
+    text: String,
+) {
+    Column {
+        Body2(
+            modifier = Modifier
+                .fillMaxWidth()
+                .align(
+                    alignment = Alignment.Start,
+                ),
+            text = text,
+            color = JobisColor.Gray600,
+        )
+        Spacer(
+            modifier = Modifier.height(10.dp),
+        )
+        Divider(
+            modifier = Modifier.fillMaxWidth(),
+            color = JobisColor.Gray400,
+        )
+    }
+}
